+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## master (unreleased)
 
+- Improve error message when background schema migration name is already taken
 - Fix copying column background migration to work with primary keys added via `initialize_column_type_change`
 
 ## 0.16.0 (2024-03-28)